Relocate romanNumeralToChordInC function
