+++ src/app/exercise/exercises/utility/exerciseAttributes/composeExercise.ts
@@ -5,7 +5,6 @@ import {
   toGetter,
 } from '../../../../shared/ts-utility';
 
-// todo: add unit tests
 // todo: consider if we can make passing "createExercise" to this, currently we had issues with typing inference
 export const composeExercise = composeWithMerge({
   defaultSettings: (value1: Exercise.Settings, value2: Exercise.Settings) => ({
