+++ src/app/exercise/utility/music/harmony/romanNumeralToChordInC.ts
@@ -1,0 +1,7 @@
+import { RomanNumeralChordSymbol } from './RomanNumeralChordSymbol';
+import { Chord } from '../chords';
+import { RomanNumeralChord } from './RomanNumeralChord';
+
+export function romanNumeralToChordInC(romanNumeralSymbol: RomanNumeralChordSymbol): Chord {
+  return new RomanNumeralChord(romanNumeralSymbol).getChord('C');
+}
